// packages/api/src/types/ArcGISConfirmedLocalAttrs.ts
export interface ArcGISConfirmedLocalAttrs {
  attributes: {
    FID: number;
    sequ: number;
    PH_masterl: string;
    edad: string;
    kasarian: string;
    nationalit: string;
    residence: string;
    travel_hx: string;
    symptoms: string;
    confirmed: string;
    facility: string;
    latitude: number;
    longitude: number;
    status: string;
    epi_link: string;
    petsa: string;
  };
}